Some Clean up for the samples
