+++ src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/KmlLayers/AddKMLLayer.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="ArcGISRuntime.Samples.Desktop.Samples.KmlLayers.AddKMLLayer"
+﻿<UserControl x:Class="ArcGISRuntime.Samples.Desktop.AddKMLLayer"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
@@ -7,14 +7,12 @@
              mc:Ignorable="d" 
              d:DesignHeight="300" d:DesignWidth="300">
     <Grid>
-
         <esri:MapView x:Name="MyMapView">
             <esri:Map>
                 <esri:ArcGISTiledMapServiceLayer ServiceUri="http://services.arcgisonline.com/arcgis/rest/services/World_Topo_Map/MapServer"/>
                 <esri:KmlLayer SourceUri="http://earthquake.usgs.gov/earthquakes/feed/v1.0/summary/1.0_week_age_link.kml"/>
             </esri:Map>
         </esri:MapView>
-
         
     </Grid>
 </UserControl>
