+++ src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/KmlLayers/AddKMLLayer.xaml.cs
@@ -1,29 +1,18 @@
-﻿using Esri.ArcGISRuntime.Layers;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
+﻿using System;
 using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
+using Esri.ArcGISRuntime.Layers;
 
-namespace ArcGISRuntime.Samples.Desktop.Samples.KmlLayers
+namespace ArcGISRuntime.Samples.Desktop
 {
     /// <summary>
-    /// This sample show you how to add online kml layer to the map
+    /// This sample demonstrates adding online kml layer to the map
     /// </summary>
     /// <title>Add online KML Layer</title>
     /// <category>Layers</category>
     /// <subcategory>Kml Layers</subcategory>
     public partial class AddKMLLayer : UserControl
     {
+        /// <summary>Construct adding online kml sample control</summary>
         public AddKMLLayer()
         {
             InitializeComponent();
