+++ src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/KmlLayers/AddLocalKMLLayer.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="ArcGISRuntime.Samples.Desktop.Samples.KmlLayers.AddLocalKMLLayer"
+﻿<UserControl x:Class="ArcGISRuntime.Samples.Desktop.AddLocalKMLLayer"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
