+++ src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/KmlLayers/AddLocalKMLLayer.xaml.cs
@@ -1,28 +1,17 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
 
-namespace ArcGISRuntime.Samples.Desktop.Samples.KmlLayers
+namespace ArcGISRuntime.Samples.Desktop
 {
     /// <summary>
-    /// This sample show you how to add local kml layer to the map
+    /// This sample demonstrates how to add local kml layer to the map
     /// </summary>
     /// <title>Add local KML Layer</title>
     /// <category>Layers</category>
     /// <subcategory>Kml Layers</subcategory>
     public partial class AddLocalKMLLayer : UserControl
     {
+        /// <summary>Construct adding local kml sample control</summary>
         public AddLocalKMLLayer()
         {
             InitializeComponent();
