+++ src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/KmlLayers/DragDropKMLFile.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="ArcGISRuntime.Samples.Desktop.Samples.KmlLayers.DragDrop_File"
+﻿<UserControl x:Class="ArcGISRuntime.Samples.Desktop.DragDropKMLFile"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
