+++ src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/KmlLayers/NetworklinkKML.xaml
@@ -1,15 +1,16 @@
-﻿<UserControl x:Class="ArcGISRuntime.Samples.Desktop.Samples.KmlLayers.NetworkLink_with_tree_view"
+﻿<UserControl x:Class="ArcGISRuntime.Samples.Desktop.NetworklinkKML"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013"
-             xmlns:converters="clr-namespace:ArcGISRuntime.Samples.Desktop.Samples.KmlLayers"
              mc:Ignorable="d" 
              d:DesignHeight="300" d:DesignWidth="300">
     <Grid>
         <Grid.Resources>
-            <converters:BooleanToVisibilityConverter x:Key="booleanToVisibilityConverter"/>
+            
+            <BooleanToVisibilityConverter x:Key="BooleanToVisibilityConverter"/>
+            
             <Storyboard x:Key="Storyboard" BeginTime="00:00:00.000" Duration="00:00:1.000" RepeatBehavior="Forever">
                 <DoubleAnimation Storyboard.TargetName="RotateImage" 
                              Storyboard.TargetProperty="(UIElement.RenderTransform).(RotateTransform.Angle)" 
@@ -23,7 +24,6 @@
                 <esri:KmlLayer SourceUri="http://radar.vlieghinder.nl/"/>
             </esri:Map>
         </esri:MapView>
-
         
         <Border x:Name="TreeViewBorder" CornerRadius="5" Background="White" BorderBrush="Black" BorderThickness="2" Margin="5,10,5,5" HorizontalAlignment="Right" VerticalAlignment="Top" Visibility="Visible">
             <StackPanel>
@@ -31,7 +31,7 @@
                     <TreeView.ItemTemplate>
                         <HierarchicalDataTemplate DataType="{x:Type MenuItem}" ItemsSource="{Binding ChildFeatures}">
                             <StackPanel Orientation="Horizontal" >
-                                <Image x:Name="RotateImage" Stretch="Uniform" Source="pack://application:,,,/ArcGISRuntimeSamplesDesktop;component/Assets/Updating.png" RenderTransformOrigin="0.5,0.5" Margin="3" Visibility="{Binding IsUpdating, FallbackValue=Collapsed, Converter={StaticResource booleanToVisibilityConverter}, Mode=OneWay}">
+                                <Image x:Name="RotateImage" Stretch="Uniform" Source="pack://application:,,,/ArcGISRuntimeSamplesDesktop;component/Assets/Updating.png" RenderTransformOrigin="0.5,0.5" Margin="3" Visibility="{Binding IsUpdating, FallbackValue=Collapsed, Converter={StaticResource BooleanToVisibilityConverter}, Mode=OneWay}">
                                     <Image.RenderTransform>
                                         <RotateTransform Angle="0" />
                                     </Image.RenderTransform>
