+++ src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/Scene/KMLLayerSample3D.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="ArcGISRuntime.Samples.Desktop.Samples.Scene.KMLLayerSample3D"
+﻿<UserControl x:Class="ArcGISRuntime.Samples.Desktop.KMLLayerSample3D"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
