+++ src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/Scene/KMLLayerSample3D.xaml.cs
@@ -1,28 +1,17 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
 
-namespace ArcGISRuntime.Samples.Desktop.Samples.Scene
+namespace ArcGISRuntime.Samples.Desktop
 {
     /// <summary>
-    /// Demonstrates adding a KML to a Scene in XAML.
+    /// This sample demonstrates adding a KML to a Scene in XAML.
     /// </summary>
     /// <title>3D KML Layer</title>
     /// <category>Scene</category>
     /// <subcategory>Layers</subcategory>
     public partial class KMLLayerSample3D : UserControl
     {
+        ///<summary>Construct 3D KML sample control</summary>  
         public KMLLayerSample3D()
         {
             InitializeComponent();
